requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4

# OCR Engines (No Tesseract / No PaddleOCR)
easyocr==1.7.1
transformers==4.40.0
torch==2.2.2
torchvision==0.17.2
timm==0.9.16

# Image Processing
opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.22.0
numpy==1.26.4
scipy==1.13.0
imutils==0.5.4

# LLM (Hackathon models: Phi-2, Qwen2-0.5B, etc.)
accelerate==0.29.3
bitsandbytes==0.43.1
sentencepiece==0.2.0

# QC Interface
streamlit==1.33.0
plotly==5.21.0
pandas==2.2.2

# Utilities
PyYAML==6.0.1
loguru==0.7.2
tqdm==4.66.2
httpx==0.27.0
aiofiles==23.2.1
python-dotenv==1.0.1

# Evaluation / metrics
scikit-learn==1.4.2
editdistance==0.8.1